Reorganize `client`'s GUI structure
