+++ client/src/app/packets_table.rs
