+++ client/src/app/rtp_streams_plot.rs
